식단 검색 위젯 대폭 개선
