+++ lib/utils/diet_util.dart
@@ -1,8 +1,13 @@
+import 'dart:convert';
+import 'dart:developer';
+import 'package:http/http.dart' as http;
+
 import 'package:bodyguard/screens/enter_calories_page/widgets/delete_diet_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 
 import '../database/config_database.dart';
+import '../model/fetched_diet_data.dart';
 
 class DietUtil {
   static final DietUtil _instance = DietUtil._internal();
@@ -12,6 +17,43 @@ class DietUtil {
   }
 
   DietUtil._internal();
+
+  /// openAPI 이용, 식품 영양 정보 가져오기
+  Future<List<FetchedDietData>> Fetchinfo(String? inputText) async{
+    if(inputText == null){
+      log('값 받아오기 실패');
+      throw Exception('Failed to load data');
+    }
+    log("받은 검색의 값${inputText}");
+    log('새로운 api');
+    var url = "http://openapi.foodsafetykorea.go.kr/api"; //sample/I2790/xml/1/5/DESC_KOR=값 &RESEARCH_YEAR=값 &MAKER_NAME=값 &FOOD_CD=값 &CHNG_DT=값";
+    url += "/e9362f2ec93a4ad2ba85";
+    url += "/I2790"; //데이터를 받는 키(키: 벨류 형식) ex) body:{}
+    url += "/json";
+    url += "/1/20/DESC_KOR=${inputText}"; //받아올 데이터의 크기 1~ 1000까지
+    log(url);
+    final uri = Uri.parse(url);
+    var response = await http.get(uri);
+    if (response.statusCode == 200) {
+      //final decoded = jsonDecode(utf8.decode(response.bodyBytes));
+      // 수정된 'items' 접근 경로
+      Map<String?, dynamic> map = await json.decode(response.body);
+      log(response.body);
+      Map<String?, dynamic> body = map["I2790"]; // 앞서 데이터를 받는 키 입력하면 됨
+      List<dynamic> item = body["row"];
+      log("리스트의 크기: ${item.length}");
+      List<FetchedDietData> allInfo = item.map((dynamic items) => FetchedDietData.fromJson(items)).toList();
+      // allInfo - Json문서를 변환
+      log('allInfo: ${allInfo.last.DESC_KOR}');
+      return allInfo;
+    } else {
+      // 오류 처리 또는 빈 리스트 반환
+      print('값 받아오기 실패');
+      log('값 받아오기 실패');
+      throw Exception('Failed to load data');
+    }
+
+  }
 
   /// 식단 정보 알림창 띄우기
   void showDietDetails(BuildContext context, DietData diet) {
